Better Images
Handles instagram, returns null if unhandled.
