+++ frontend/utils.tsx
@@ -35,3 +35,27 @@ export const readTBA = async (url: string) => {
 
   throw new Error("TBA Error: " + response.status);
 };
+
+export const getMediaUrl = async (team: number, year: number) => {
+  const data = await readTBA(`/team/frc${team}/media/${year}`);
+  const image = data.filter((item: any) => item?.preferred)?.[0];
+  if (image?.type === "instagram-image") {
+    if (image?.view_url) {
+      return `https://www.thebluealliance.com/${image?.direct_url}`;
+    }
+    return null;
+  } else if (image?.type === "imgur") {
+    return image?.direct_url ?? null;
+  } else {
+    return null;
+  }
+};
+
+export const getMediaUrls = async (teams: number[], year: number) => {
+  const urls = [];
+  for (const team of teams) {
+    const url = await getMediaUrl(team, year);
+    urls.push(url);
+  }
+  return urls;
+};
